# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Go4Ride API",
    "description": "\nRide-hailing **rider app** API (Phase 0\u20132). Driver and admin APIs are out of scope for this service.\n\n## Authentication\n\n1. `POST /api/v1/auth/request-otp` \u2014 send OTP to phone (`debug_otp` returned when `OTP_DEBUG=true`)\n2. `POST /api/v1/auth/verify-otp` \u2014 exchange OTP for JWT pair; creates account on first sign-in\n3. Click **Authorize** and paste the `access_token` (Bearer, no prefix)\n4. `POST /api/v1/auth/refresh` \u2014 new token pair using `refresh_token` in body (no header)\n\n## Typical booking flow\n\n1. Reverse geocode (optional) \u2192 list ride types \u2192 estimate fare\n2. `POST /api/v1/rides` with optional `Idempotency-Key` header\n3. `WS /api/v1/ws/rides/{ride_id}?token={access_token}` for live updates\n4. Cancel, history, repeat, or fetch invoice as needed\n\n## WebSocket\n\n```\nWS /api/v1/ws/rides/{ride_id}?token=<access_token>\n```\n\nEvents are JSON payloads published on ride status changes (mock driver advances stages in dev).\n\n## Errors\n\nStructured JSON errors: `{\"detail\": \"...\", \"code\": \"...\"}` with HTTP 4xx/5xx.\n",
    "version": "0.2.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development"
    },
    {
      "url": "https://go4ride-api.onrender.com",
      "description": "Render (example production)"
    }
  ],
  "paths": {
...
    "/api/v1/rides/estimate": {
      "post": {
        "tags": [
          "rides"
        ],
        "summary": "Estimate Ride",
        "description": "Get distance, duration, and fare quote for a route.",
        "operationId": "estimate_ride_api_v1_rides_estimate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RideEstimateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RideEstimateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rides": {
      "post": {
        "tags": [
          "rides"
        ],
        "summary": "Create Ride",
        "description": "Book a ride. Optional `Idempotency-Key` header prevents duplicate bookings.",
        "operationId": "create_ride_api_v1_rides_post",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRideRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RideResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rides/history": {
      "get": {
        "tags": [
          "rides"
        ],
        "summary": "Ride History",
        "description": "Paginated ride history for the Bookings screen.",
        "operationId": "ride_history_api_v1_rides_history_get",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "terminal (default), all, completed, or cancelled",
              "default": "terminal",
              "title": "Status"
            },
            "description": "terminal (default), all, completed, or cancelled"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RideHistoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rides/{ride_id}/repeat": {
      "post": {
        "tags": [
          "rides"
        ],
        "summary": "Repeat Ride",
        "description": "Return estimate + create payload to re-book a past ride.",
        "operationId": "repeat_ride_api_v1_rides__ride_id__repeat_post",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "ride_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Ride Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepeatRideResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rides/{ride_id}/cancel": {
      "post": {
        "tags": [
          "rides"
        ],
        "summary": "Cancel Ride",
        "description": "Cancel an active ride owned by the current rider.",
        "operationId": "cancel_ride_api_v1_rides__ride_id__cancel_post",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "ride_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Ride Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RideResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rides/{ride_id}/status": {
      "get": {
        "tags": [
          "rides"
        ],
        "summary": "Ride Status",
        "description": "Lightweight status poll (prefer WebSocket for live updates).",
        "operationId": "ride_status_api_v1_rides__ride_id__status_get",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "ride_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Ride Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RideStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rides/{ride_id}": {
      "get": {
        "tags": [
          "rides"
        ],
        "summary": "Get Ride",
        "description": "Full ride details including driver info when assigned.",
        "operationId": "get_ride_api_v1_rides__ride_id__get",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "ride_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Ride Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RideResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rides/{ride_id}/invoice": {
      "get": {
        "tags": [
          "rides"
        ],
        "summary": "Ride Invoice",
        "description": "Receipt / invoice stub for completed rides.",
        "operationId": "ride_invoice_api_v1_rides__ride_id__invoice_get",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "ride_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Ride Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/settings": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "Get Settings",
        "description": "Get notification and app preferences.",
        "operationId": "get_settings_api_v1_settings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "settings"
        ],
        "summary": "Update Settings",
        "description": "Update notification and app preferences.",
        "operationId": "update_settings_api_v1_settings_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Coordinates-Input": {
        "properties": {
          "lat": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 90.0,
                "minimum": -90.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Lat"
          },
          "lng": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 180.0,
                "minimum": -180.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Lng"
          }
        },
        "type": "object",
        "required": [
          "lat",
          "lng"
        ],
        "title": "Coordinates"
      },
...
      "CreateRideRequest": {
        "properties": {
          "pickup": {
            "$ref": "#/components/schemas/Coordinates-Input"
          },
          "drop": {
            "$ref": "#/components/schemas/Coordinates-Input"
          },
          "pickup_address": {
            "type": "string",
            "title": "Pickup Address"
          },
          "drop_address": {
            "type": "string",
            "title": "Drop Address"
          },
          "ride_type_slug": {
            "type": "string",
            "title": "Ride Type Slug"
          }
        },
        "type": "object",
        "required": [
          "pickup",
          "drop",
          "pickup_address",
          "drop_address",
          "ride_type_slug"
        ],
        "title": "CreateRideRequest",
        "examples": [
          {
            "drop": {
              "lat": "12.9352",
              "lng": "77.6245"
            },
            "drop_address": "Koramangala, Bengaluru",
            "pickup": {
              "lat": "12.9716",
              "lng": "77.5946"
            },
            "pickup_address": "MG Road, Bengaluru",
            "ride_type_slug": "mini"
          }
        ]
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InvoiceResponse": {
        "properties": {
          "available": {
            "type": "boolean",
            "title": "Available"
          },
          "ride_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ride Id"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "pickup_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pickup Address"
          },
          "drop_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Drop Address"
          },
          "final_fare": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Fare"
          },
          "currency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currency"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "driver": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DriverSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "download_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Download Url"
          }
        },
        "type": "object",
        "required": [
          "available"
        ],
        "title": "InvoiceResponse"
      },
...
      "RepeatRideResponse": {
        "properties": {
          "pickup": {
            "$ref": "#/components/schemas/Coordinates-Output"
          },
          "drop": {
            "$ref": "#/components/schemas/Coordinates-Output"
          },
          "pickup_address": {
            "type": "string",
            "title": "Pickup Address"
          },
          "drop_address": {
            "type": "string",
            "title": "Drop Address"
          },
          "ride_type_slug": {
            "type": "string",
            "title": "Ride Type Slug"
          }
        },
        "type": "object",
        "required": [
          "pickup",
          "drop",
          "pickup_address",
          "drop_address",
          "ride_type_slug"
        ],
        "title": "RepeatRideResponse"
      },
...
      "RideEstimateRequest": {
        "properties": {
          "pickup": {
            "$ref": "#/components/schemas/Coordinates-Input"
          },
          "drop": {
            "$ref": "#/components/schemas/Coordinates-Input"
          },
          "ride_type_slug": {
            "type": "string",
            "title": "Ride Type Slug",
            "examples": [
              "mini",
              "sedan",
              "bike",
              "xl"
            ]
          }
        },
        "type": "object",
        "required": [
          "pickup",
          "drop",
          "ride_type_slug"
        ],
        "title": "RideEstimateRequest",
        "examples": [
          {
            "drop": {
              "lat": "12.9352",
              "lng": "77.6245"
            },
            "pickup": {
              "lat": "12.9716",
              "lng": "77.5946"
            },
            "ride_type_slug": "mini"
          }
        ]
      },
      "RideEstimateResponse": {
        "properties": {
          "distance_km": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Distance Km"
          },
          "duration_min": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Duration Min"
          },
          "estimated_fare": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Estimated Fare"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "surge_multiplier": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Surge Multiplier"
          }
        },
        "type": "object",
        "required": [
          "distance_km",
          "duration_min",
          "estimated_fare",
          "currency",
          "surge_multiplier"
        ],
        "title": "RideEstimateResponse"
      },
      "RideHistoryResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RideResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "page",
          "limit",
          "total"
        ],
        "title": "RideHistoryResponse"
      },
      "RideResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "pickup_lat": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Pickup Lat"
          },
          "pickup_lng": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Pickup Lng"
          },
          "pickup_address": {
            "type": "string",
            "title": "Pickup Address"
          },
          "drop_lat": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Drop Lat"
          },
          "drop_lng": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Drop Lng"
          },
          "drop_address": {
            "type": "string",
            "title": "Drop Address"
          },
          "estimated_fare": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Estimated Fare"
          },
          "final_fare": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Fare"
          },
          "distance_km": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance Km"
          },
          "duration_min": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Min"
          },
          "surge_multiplier": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Surge Multiplier"
          },
          "ride_type_slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ride Type Slug"
          },
          "requested_at": {
            "type": "string",
            "format": "date-time",
            "title": "Requested At"
          },
          "driver_assigned_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Driver Assigned At"
          },
          "driver_arrived_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Driver Arrived At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "cancelled_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancelled At"
          },
          "driver": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DriverSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "invoice_available": {
            "type": "boolean",
            "title": "Invoice Available",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "pickup_lat",
          "pickup_lng",
          "pickup_address",
          "drop_lat",
          "drop_lng",
          "drop_address",
          "estimated_fare",
          "final_fare",
          "distance_km",
          "duration_min",
          "surge_multiplier",
          "requested_at"
        ],
        "title": "RideResponse"
      },
      "RideStatusResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "driver": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DriverSummary"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "status"
        ],
        "title": "RideStatusResponse"
      },
...
      "SettingsResponse": {
        "properties": {
          "notifications_enabled": {
            "type": "boolean",
            "title": "Notifications Enabled"
          },
          "language": {
            "type": "string",
            "title": "Language"
          }
        },
        "type": "object",
        "required": [
          "notifications_enabled",
          "language"
        ],
        "title": "SettingsResponse"
      },
      "SettingsUpdateRequest": {
        "properties": {
          "notifications_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notifications Enabled"
          },
          "language": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 8,
                "minLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          }
        },
        "type": "object",
        "title": "SettingsUpdateRequest"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "JWT access token from `POST /api/v1/auth/verify-otp`. Paste the token only (Swagger adds the Bearer prefix)."
      }
    }
  },
...
}